reorganize makefile modules
